+++ CHANGELOG.md
@@ -12,3 +12,5 @@ v0.12.0-alpha23
 #### Bug fixes
 
 * Added kbox.app.exists and kbox.app.nextAppName in [#1235](https://github.com/kalabox/kalabox/issues/1235) for use in [#1258](https://github.com/kalabox/kalabox/issues/1258)
+
+* Changed syncthing rescan interval to always be 2 seconds. [#1199](https://github.com/kalabox/kalabox/issues/1199)